Asset fee sponsorship (#127)
* State API: expose accounts data storage methods

* Accounts Data Storage: make immediate entries updates

When Data tx is performed, it is expected that entries appear in state immediately, since
they can be used by following transactions even in same block.
In old implementation entries are written to DB batch, appearing in state only after this batch
is written.
In new implementation entries are written to local storage (and returned from there) and at the end
of applying blocks batch are written to DB batch.

* Implement sponsored assets storage and related functions

* TestHeadersValidation: do not modify cached blocks slice

Otherwise it results in failure in further tests.

* Sponsorship: min fee validation

*  Few fixes for sponsored assets storage

 - Use `math/big` for fromWaves/toWaves calculations (like in Scala implementation).
 - If asset's cost is 0, don't consider it sponsored anymore.

* Split State interface

* Tiny refactoring in account's data storage tests

* Tests for block differ

* Support sponsorship in NG fees distribution

* createDiffTransfer: support fee sponsorship

* Block differ: tests for sponsorship

* CheckTransfer(): tests for sponsorship

* TestCreateDiffTransfer(): add tests for sponsorship

* Validation of Sponsorship transaction

* Small fix

* Allow Sponsorship transactions directly after feature activation

This is the way it is in real blockchain.
Sponsorship transactions were allowed but sponsorship started to work later.

* Do not check Waves as fee asset

* Fix Fair PoS base target calculation

* Exchange tx: only use pointers for Orders (for all the versions)

Before that, both pointer and value types for Order implemented Order interface.
But when casting in OrderToOrderBody and other functions, only value was accepted.
This leads to errors when creating exchange with pointer, which is also valid Order.
The good solution is to fixate that only pointer implements Order interface.

* State: global tests refactoring

* Fix typo

+++ pkg/consensus/pos_calculator.go
@@ -152,10 +152,10 @@ func (calc *FairPosCalculator) CalculateBaseTarget(
 	if greatGrandParentTimestamp == 0 {
 		return confirmedTarget, nil
 	}
-	average := float64(applyingBlockTimestamp-greatGrandParentTimestamp) / 3 / 1000
-	if average > maxDelay {
+	average := (applyingBlockTimestamp - greatGrandParentTimestamp) / 3 / 1000
+	if float64(average) > maxDelay {
 		return (confirmedTarget + uint64(math.Max(1, float64(confirmedTarget/100)))), nil
-	} else if average < minDelay {
+	} else if float64(average) < minDelay {
 		return (confirmedTarget - uint64(math.Max(1, float64(confirmedTarget/100)))), nil
 	} else {
 		return confirmedTarget, nil
